+++ test/ResourceGeneratorTest.php
@@ -10,9 +10,13 @@ namespace ZendTest\Expressive\Hal;
 use ArrayIterator;
 use Generator;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use stdClass;
 use Zend\Expressive\Hal\Exception\InvalidObjectException;
+use Zend\Expressive\Hal\Exception\InvalidStrategyException;
+use Zend\Expressive\Hal\Exception\UnknownMetadataTypeException;
 use Zend\Expressive\Hal\HalResource;
 use Zend\Expressive\Hal\Link;
 use Zend\Expressive\Hal\LinkGenerator;
@@ -22,6 +26,7 @@ use Zend\Expressive\Hal\ResourceGenerator\Exception\OutOfBoundsException;
 use Zend\Hydrator\ObjectProperty as ObjectPropertyHydrator;
 use Zend\Paginator\Adapter\ArrayAdapter;
 use Zend\Paginator\Paginator;
+use ZendTest\Expressive\Hal\TestAsset\TestMetadata;
 
 /**
  * @todo Create tests for cases where resources embed other resources.
@@ -29,6 +34,31 @@ use Zend\Paginator\Paginator;
 class ResourceGeneratorTest extends TestCase
 {
     use Assertions;
+
+    /**
+     * @var ObjectProphecy|ServerRequestInterface
+     */
+    private $request;
+
+    /**
+     * @var ObjectProphecy|ContainerInterface
+     */
+    private $hydrators;
+
+    /**
+     * @var ObjectProphecy|LinkGenerator
+     */
+    private $linkGenerator;
+
+    /**
+     * @var ObjectProphecy|Metadata\MetadataMap
+     */
+    private $metadataMap;
+
+    /**
+     * @var ObjectProphecy|ResourceGenerator
+     */
+    private $generator;
 
     public function setUp()
     {
@@ -40,6 +70,26 @@ class ResourceGeneratorTest extends TestCase
             $this->metadataMap->reveal(),
             $this->hydrators->reveal(),
             $this->linkGenerator->reveal()
+        );
+
+        $this->generator->addStrategy(
+            Metadata\RouteBasedResourceMetadata::class,
+            ResourceGenerator\RouteBasedResourceStrategy::class
+        );
+
+        $this->generator->addStrategy(
+            Metadata\RouteBasedCollectionMetadata::class,
+            ResourceGenerator\RouteBasedCollectionStrategy::class
+        );
+
+        $this->generator->addStrategy(
+            Metadata\UrlBasedCollectionMetadata::class,
+            ResourceGenerator\UrlBasedCollectionStrategy::class
+        );
+
+        $this->generator->addStrategy(
+            Metadata\UrlBasedResourceMetadata::class,
+            ResourceGenerator\UrlBasedResourceStrategy::class
         );
     }
 
@@ -682,5 +732,19 @@ class ResourceGeneratorTest extends TestCase
         $this->expectException(ResourceGenerator\Exception\InvalidCollectionException::class);
         $this->expectExceptionMessage('not a Traversable');
         $this->generator->fromObject($instance, $this->request->reveal());
+    }
+
+    public function testAddStrategyRaisesExceptionIfInvalidMetadataClass()
+    {
+        $this->expectException(UnknownMetadataTypeException::class);
+        $this->expectExceptionMessage('does not exist, or does not extend');
+        $this->generator->addStrategy(stdClass::class, 'invalid-strategy');
+    }
+
+    public function testAddStrategyRaisesExceptionIfInvalidStrategyClass()
+    {
+        $this->expectException(InvalidStrategyException::class);
+        $this->expectExceptionMessage('does not exist, or does not implement');
+        $this->generator->addStrategy(TestMetadata::class, 'invalid-strategy');
     }
 }