+++ test/TestAsset/TestMetadataMapFactory.php
@@ -7,13 +7,12 @@
 
 namespace ZendTest\Expressive\Hal\TestAsset;
 
-use stdClass;
-use Zend\Expressive\Hal\Metadata\AbstractMetadata;
+use Zend\Expressive\Hal\Metadata\MetadataMapFactory;
 
-class TestMetadata extends AbstractMetadata
+class TestMetadataMapFactory extends MetadataMapFactory
 {
-    public function getClass() : string
+    protected function createTestMetadata(array $metadata) : TestMetadata
     {
-        return stdClass::class;
+        return new TestMetadata();
     }
 }